add published status to timeline events
To publish, add "published" attrib to root tag:
    <Timeline published="1">

This replaces the previous commit. Instead of not emitting any
events if timeline.xml not marked published, determine status
and then add published attribute to each event.
This allows front to show or hide the events page based on
settings.DEBUG, so we can check stuff in stage.
refs #212, refs #215